integracion/python/requirements.txt
psycopg2-binary==2.9.9
pymysql==1.1.1
cryptography>=41.0.0
